mock user location in test super class
